rename personalized approach to cot learner profile
