+++ athena/modules/text/module_text_llm/module_text_llm/config.py
@@ -6,10 +6,10 @@ from athena import config_schema_provider
 from module_text_llm.chain_of_thought_approach import ChainOfThoughtConfig
 from module_text_llm.basic_approach import BasicApproachConfig
 from module_text_llm.divide_and_conquer import DivideAndConquerConfig
-from module_text_llm.personalized_approach import PersonalizedConfig
+from module_text_llm.cot_learner_profile import COTLearnerProfileConfig
 from module_text_llm.llm_as_profiler import LLMAsProfilerConfig
 
-ApproachConfigUnion = Union[BasicApproachConfig, ChainOfThoughtConfig, DivideAndConquerConfig, SelfConsistencyConfig, PersonalizedConfig, LLMAsProfilerConfig]
+ApproachConfigUnion = Union[BasicApproachConfig, ChainOfThoughtConfig, DivideAndConquerConfig, SelfConsistencyConfig, COTLearnerProfileConfig, LLMAsProfilerConfig]
 
 @config_schema_provider
 class Configuration(BaseModel):
