+++ athena/modules/text/module_text_llm/module_text_llm/cot_learner_profile/__init__.py
@@ -3,14 +3,14 @@ from typing import Literal
 from athena.text import Exercise, Submission
 
 from module_text_llm.approach_config import ApproachConfig
-from module_text_llm.personalized_approach.prompt_generate_feedback import GenerateSuggestionsPrompt
-from module_text_llm.personalized_approach.prompt_thinking import ThinkingPrompt
-from module_text_llm.personalized_approach.generate_suggestions import generate_suggestions
+from module_text_llm.cot_learner_profile.prompt_generate_feedback import GenerateSuggestionsPrompt
+from module_text_llm.cot_learner_profile.prompt_thinking import ThinkingPrompt
+from module_text_llm.cot_learner_profile.generate_suggestions import generate_suggestions
 from athena.schemas.learner_profile import LearnerProfile
 
 
-class PersonalizedConfig(ApproachConfig):
-    type: Literal['personalized'] = 'personalized'
+class COTLearnerProfileConfig(ApproachConfig):
+    type: Literal['cot_learner_profile'] = 'cot_learner_profile'
     thinking_prompt: ThinkingPrompt = Field(default=ThinkingPrompt())
     generate_suggestions_prompt: GenerateSuggestionsPrompt = Field(default=GenerateSuggestionsPrompt())
