+++ athena/modules/text/module_text_llm/module_text_llm/cot_learner_profile/prompt_thinking.py
